refac: extract tools to a separate package
