+++ viewer/core/__testutilities__.ts
@@ -1,0 +1,4 @@
+/*!
+ * Copyright 2021 Cognite AS
+ */
+export * from './src/__testutilities__';
