+++ viewer/core/cad.ts
@@ -1,0 +1,5 @@
+/*!
+ * Copyright 2021 Cognite AS
+ */
+export { LevelOfDetail } from './src/datamodels/cad/sector/LevelOfDetail';
+export { SectorNode } from './src/datamodels/cad/sector/SectorNode';
