+++ viewer/core/index.ts
@@ -1,0 +1,7 @@
+/*!
+ * Copyright 2021 Cognite AS
+ */
+export { Cognite3DViewer } from './src/public/migration/Cognite3DViewer';
+export { Cognite3DModel } from './src/public/migration/Cognite3DModel';
+export { CogniteModelBase } from './src/public/migration/CogniteModelBase';
+export * from './src/public/types';
