+++ viewer/core/internals.ts
@@ -1,5 +1,4 @@
 /*!
  * Copyright 2021 Cognite AS
  */
-
-export * from './core/internals';
+export * from './src/internals';
