--- viewer/core/src/tools.ts
@@ -1,22 +1,0 @@
-/*!
- * Copyright 2021 Cognite AS
- */
-
-/**
- * @module @cognite/reveal/tools
- */
-
-export { HtmlOverlayTool, HtmlOverlayOptions } from './tools/HtmlOverlayTool';
-export { ExplodedViewTool } from './tools/ExplodedViewTool';
-export { DebugCameraTool } from './tools/DebugCameraTool';
-export { AxisViewTool } from './tools/AxisView/AxisViewTool';
-export {
-  AxisBoxConfig,
-  AxisBoxCompassConfig,
-  AxisBoxFaceConfig,
-  Corner,
-  AbsolutePosition,
-  RelativePosition
-} from './tools/AxisView/types';
-export { Cognite3DViewerToolBase } from './tools/Cognite3DViewerToolBase';
-export { DebugLoadedSectorsTool, DebugLoadedSectorsToolOptions } from './tools/DebugLoadedSectorsTool';
