+++ viewer/packages/tools/index.ts
@@ -1,0 +1,18 @@
+/*!
+ * Copyright 2021 Cognite AS
+ */
+
+export { HtmlOverlayTool, HtmlOverlayOptions } from './src/HtmlOverlayTool';
+export { ExplodedViewTool } from './src/ExplodedViewTool';
+export { DebugCameraTool } from './src/DebugCameraTool';
+export { AxisViewTool } from './src/AxisView/AxisViewTool';
+export {
+  AxisBoxConfig,
+  AxisBoxCompassConfig,
+  AxisBoxFaceConfig,
+  Corner,
+  AbsolutePosition,
+  RelativePosition
+} from './src/AxisView/types';
+export { Cognite3DViewerToolBase } from './src/Cognite3DViewerToolBase';
+export { DebugLoadedSectorsTool, DebugLoadedSectorsToolOptions } from './src/DebugLoadedSectorsTool';
