+++ viewer/packages/tools/src/Cognite3DViewerToolBase.test.ts
@@ -2,7 +2,7 @@
  * Copyright 2021 Cognite AS
  */
 
-import { DisposedDelegate } from '../public/types';
+import { DisposedDelegate } from './types';
 import { Cognite3DViewerToolBase } from './Cognite3DViewerToolBase';
 
 describe('Cognite3DViewerToolBase', () => {
