+++ viewer/packages/tools/src/HtmlOverlayTool.ts
@@ -3,10 +3,8 @@
  */
 
 import * as THREE from 'three';
-import { Cognite3DViewer } from '../public/migration/Cognite3DViewer';
-import { DisposedDelegate, SceneRenderedDelegate } from '../public/types';
-
-import { worldToViewport } from '../utilities/worldToViewport';
+import { Cognite3DViewer, DisposedDelegate, SceneRenderedDelegate } from '@reveal/core';
+import { worldToViewport } from '@reveal/core/utilities';
 import { Cognite3DViewerToolBase } from './Cognite3DViewerToolBase';
 
 export type HtmlOverlayPositionUpdatedDelegate = (
