+++ viewer/packages/tools/src/types.ts
@@ -1,0 +1,8 @@
+/*!
+ * Copyright 2021 Cognite AS
+ */
+
+/*!
+ * Copyright 2021 Cognite AS
+ */
+export type DisposedDelegate = () => void;
